fix(express): render every /components/<slug> page and a branded 404
Cloudflare Pages only had 9 top-level routes to fall back to, so all 30
component detail links served the home page with HTTP 200 (soft 404) —
the component catalog's primary purpose was dead on production. There
was also no 404.html at all, so every unknown path (typos, /__live,
/404.html itself) served home with 200, masking the bug above and
handing search engines duplicate content.

Route rendering (the layout substitution, the route table, the
per-component page context, and the new not-found page) is factored out
of server.js into page.js so both the Express dev server and
build-pages.js render identical HTML from one place. build-pages.js now
writes a dist/<slug>/ page for every catalog entry plus dist/404.html
(which Cloudflare Pages serves with a real 404 status once the file
exists), and server.js's catch-all now answers unknown paths with the
same 404 page instead of Express's default text response. Per-component
demo scripts move to component-demo-scripts.js, shared the same way.

The static build also stops shipping the dev-only live-reload snippet
(`new EventSource('/__live')`), which has no route on Pages and logged
a console error on every page load.

Threads an `hasApi` flag through the route table (true from the dev
server, false from the static build) for the /tasks fix in a following
commit.
