+++ examples/express/page.js
@@ -1,0 +1,223 @@
+/**
+ * Site page renderer shared by the Express dev server (server.js) and the
+ * static export (scripts/build-pages.js). Both hosts must emit the same HTML
+ * for a route, so the layout substitution, the route table, the per-component
+ * page context and the not-found page live here rather than in either caller.
+ */
+import { readFileSync } from 'node:fs';
+import { dirname, join } from 'node:path';
+import { fileURLToPath } from 'node:url';
+import { render } from '@basenative/server';
+import { renderBreadcrumb } from '../../packages/components/src/index.js';
+import {
+  getComponentsPageContext,
+  getHomePageContext,
+  getRoadmapPageContext,
+  getTasksPageContext,
+  navPages,
+} from './site-data.js';
+import { componentCategories, flatComponents, findComponent } from './component-catalog.js';
+import { getDemo } from './component-demos.js';
+import { getDemoScripts } from './component-demo-scripts.js';
+import { getShowcaseContext } from './showcase-data.js';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+const read = (file) => readFileSync(join(__dirname, file), 'utf-8');
+
+export const SITE_URL = 'https://basenative.com';
+export const DEFAULT_DESCRIPTION =
+  'BaseNative — a signal-based runtime over native HTML. Zero build step. Zero deps in core. Semantic by construction.';
+
+export function escapeAttr(s) {
+  return String(s).replace(/&/g, '&amp;').replace(/"/g, '&quot;').replace(/</g, '&lt;');
+}
+
+/**
+ * Render a view through views/layout.html.
+ *
+ * @param {string} viewFile  File name under views/
+ * @param {object} ctx       Template context for the view
+ * @param {object} opts
+ * @param {string} opts.title
+ * @param {string} [opts.description]
+ * @param {string} [opts.ogTitle]
+ * @param {string} [opts.path]        Site-relative path for canonical/og:url
+ * @param {string} [opts.scripts]     Extra markup injected before </body>
+ * @param {string} [opts.activePage]  navPages entry to mark aria-current
+ */
+export function renderPage(viewFile, ctx, opts) {
+  const {
+    title,
+    description = DEFAULT_DESCRIPTION,
+    ogTitle = `${title} — BaseNative`,
+    path = '/',
+    scripts = '',
+    activePage = '',
+  } = opts;
+
+  const layout = read('views/layout.html');
+  const view = read(`views/${viewFile}`);
+  const content = render(view, ctx);
+  let html = layout
+    .replace(/<!--TITLE-->/g, escapeAttr(title))
+    .replace(/<!--DESCRIPTION-->/g, escapeAttr(description))
+    .replace(/<!--OG_TITLE-->/g, escapeAttr(ogTitle))
+    .replace(/<!--OG_URL-->/g, escapeAttr(SITE_URL + path))
+    .replace('<!--CONTENT-->', content)
+    .replace('<!--SCRIPTS-->', scripts);
+  for (const page of navPages) {
+    html = html.replace(
+      `<!--${page.toUpperCase()}_ARIA-->`,
+      activePage === page ? 'aria-current="page"' : '',
+    );
+  }
+  return html;
+}
+
+const TEST_SIGNAL_ITEMS = [
+  { id: 1, name: 'Server-rendered item A', status: 'done' },
+  { id: 2, name: 'Server-rendered item B', status: 'active' },
+  { id: 3, name: 'Server-rendered item C', status: 'pending' },
+];
+
+/**
+ * Top-level routes. `ctx(site)` receives host-specific state — today only the
+ * task list, which is an in-memory store on the dev server and a fixed
+ * snapshot in the static export.
+ */
+export const siteRoutes = [
+  {
+    path: '/',
+    view: 'home.html',
+    title: 'Home',
+    description: DEFAULT_DESCRIPTION,
+    ogTitle: 'BaseNative — Semantic HTML + Signals',
+    activePage: 'home',
+    ctx: () => getHomePageContext(),
+  },
+  {
+    path: '/tasks',
+    view: 'tasks.html',
+    title: 'Tasks',
+    description: 'Live task list demonstrating signal-based hydration over server-rendered HTML.',
+    activePage: 'tasks',
+    ctx: ({ tasks, hasApi }) => getTasksPageContext(tasks, hasApi),
+  },
+  {
+    path: '/playground',
+    view: 'playground.html',
+    title: 'Playground',
+    description:
+      'Interactive sandbox for signals, computed values, effects, and template directives.',
+    activePage: 'playground',
+    ctx: () => ({}),
+  },
+  {
+    path: '/builder',
+    view: 'builder.html',
+    title: 'Builder',
+    description:
+      'Drag-and-drop visual builder powered by @basenative/builder — compose a layout and export BaseNative code.',
+    activePage: 'builder',
+    ctx: () => ({}),
+  },
+  {
+    path: '/docs',
+    view: 'docs.html',
+    title: 'API Docs',
+    description:
+      'API reference for @basenative/runtime, @basenative/server, @basenative/router, @basenative/forms, and @basenative/components.',
+    activePage: 'docs',
+    ctx: () => ({}),
+  },
+  {
+    path: '/components',
+    view: 'components.html',
+    title: 'Components',
+    description: `${flatComponents.length} semantic, server-rendered components built on native HTML primitives. No virtual DOM.`,
+    ogTitle: 'BaseNative Components — Semantic by Construction',
+    activePage: 'components',
+    ctx: () => getComponentsPageContext(),
+  },
+  {
+    path: '/roadmap',
+    view: 'roadmap.html',
+    title: 'Roadmap',
+    description:
+      'BaseNative release plan, trust blockers, browser policy, and workflow parity tracking.',
+    activePage: 'roadmap',
+    ctx: () => getRoadmapPageContext(),
+  },
+  {
+    path: '/test-signals',
+    view: 'test-signals.html',
+    title: 'Signal Verification',
+    activePage: '',
+    ctx: () => ({ items: TEST_SIGNAL_ITEMS, itemsJson: JSON.stringify(TEST_SIGNAL_ITEMS) }),
+  },
+  {
+    path: '/showcase',
+    view: 'showcase.html',
+    title: 'Showcase',
+    description: `Live gallery of all ${flatComponents.length} BaseNative components — every section is a real server render, not a mockup.`,
+    ogTitle: 'BaseNative Showcase — Live Component Gallery',
+    activePage: 'showcase',
+    ctx: () => ({ ...getShowcaseContext(), categories: componentCategories }),
+  },
+];
+
+export function renderRoute(route, site = {}) {
+  return renderPage(route.view, route.ctx(site), route);
+}
+
+/** Render /components/<slug>; returns null for a slug the catalog does not list. */
+export function renderComponentPage(slug) {
+  const component = findComponent(slug);
+  if (!component) return null;
+
+  const demo = getDemo(component.slug);
+  const idx = flatComponents.findIndex((c) => c.slug === component.slug);
+  const prev = idx > 0 ? flatComponents[idx - 1] : null;
+  const next = idx < flatComponents.length - 1 ? flatComponents[idx + 1] : null;
+  const related = flatComponents
+    .filter((c) => c.categoryId === component.categoryId && c.slug !== component.slug)
+    .slice(0, 6);
+
+  const ctx = {
+    component,
+    quickstart: demo?.quickstart ?? '',
+    examples: demo?.examples ?? [],
+    prev,
+    next,
+    related,
+    breadcrumb: renderBreadcrumb({
+      items: [
+        { label: 'Home', href: '/' },
+        { label: 'Components', href: '/components' },
+        { label: component.title },
+      ],
+    }),
+  };
+
+  return renderPage('component.html', ctx, {
+    title: component.title,
+    description: component.summary,
+    ogTitle: `${component.title} — BaseNative`,
+    path: `/components/${component.slug}`,
+    activePage: 'components',
+    scripts: getDemoScripts(component.slug),
+  });
+}
+
+/** The 404 page: served by the dev server for unknown paths and written to dist/404.html for Cloudflare Pages. */
+export function renderNotFoundPage() {
+  return renderPage(
+    'not-found.html',
+    {},
+    {
+      title: 'Page not found',
+      description: 'There is nothing at this address on basenative.com.',
+      path: '/404.html',
+    },
+  );
+}
