+++ examples/express/views/not-found.html
@@ -1,0 +1,13 @@
+<section data-bn-page-hero data-bn-not-found>
+  <p data-bn-eyebrow>404</p>
+  <h2>Page <em>not found</em></h2>
+  <p data-bn-lede>
+    There is nothing at this address. The link may be out of date, or the URL may have a typo —
+    every page on the site is reachable from the navigation above.
+  </p>
+  <nav data-bn-page-hero-actions aria-label="Where to go next">
+    <a href="/" data-bn-cta="primary">Back to home →</a>
+    <a href="/components" data-bn-cta="secondary">Browse components</a>
+    <a href="/docs" data-bn-cta="secondary">Read the API docs</a>
+  </nav>
+</section>
